Both PATT's internal (R,S) shelf simulation and the DES port move from
discrete integer units to continuous quantities:

- the integer rounding of delivered/order quantities added in 3cc93a6 is
  reverted, and the demand carry-over discretization is removed
- the per-customer Bernoulli FIFO/LIFO draw becomes a fluid split
  (d_fifo = P_FIFO * demand, remainder LIFO), with the FIFO stream served
  first under scarcity - mirrored identically on both sides
- DES shelves are now batch lists [[qty, expiry], ...] rather than one entry
  per unit, with _batch_qty() / _remove_from_back() helpers
- DROP_THRESHOLD 5 -> 2.0, appropriate for continuous units
- a delivered_u conservation figure (demand - stockout + waste) is reported
- validate_anylogic.py parses the AnyLogic CSVs as float

No conflict with the web-tool layer: DESPort, StoreCfg and kpis_to_row keep
their signatures, so webtool/pipeline.py drives the new simulation unchanged.


# Conflicts:
#	.gitignore

+++ validate_anylogic.py
@@ -10,21 +10,21 @@ def run(folder, suffix, start_date="2025-03-03", warmup_days=14, weeks=52.0):
     for row in csv.reader(open(p("planned_demands")), delimiter=";"):
         if len(row) < 5 or float(row[0]) < cutoff_h:
             continue
-        demA += int(row[2]); demB += int(row[3]); demC += int(row[4])
+        demA += float(row[2]); demB += float(row[3]); demC += float(row[4])
 
     wA = wB = wC = 0; co2fw = 0.0
     r = csv.reader(open(p("WeeklyWasteReport")), delimiter=";"); next(r)
     for row in r:
         if not row or not row[0].isdigit() or int(row[0]) <= warmup_days // 7:
             continue
-        wA += int(row[4]); wB += int(row[5]); wC += int(row[6]); co2fw += float(row[7])
+        wA += float(row[4]); wB += float(row[5]); wC += float(row[6]); co2fw += float(row[7])
 
     lost = 0; locost = 0.0
     r = csv.reader(open(p("WeeklyStockoutReport")), delimiter=";"); next(r)
     for row in r:
         if not row or not row[0].isdigit() or int(row[0]) <= warmup_days // 7:
             continue
-        lost += int(row[3]); locost += float(row[7])
+        lost += float(row[3]); locost += float(row[7])
 
     km = fuel = co2g = 0.0; ntr = 0
     for row in csv.reader(open(p("trucks")), delimiter=";"):
